Added multiple names for "range"
In loading data in the dataprep module, we previously
expected a single range field in the dive .mat files.
Instead, as we attempt more range calculation methods,
some of the names indicate the calculation method.

With this commit, we have a clumsy method to identify
appropraite range measurements to load into ddat