Changed order of z_adcp
Had a severe issue with selecting adcp measurements.  In
matbuilder.py, functions `get_zadcp()` and `zadcp_select()` use
a numpy array, sliced with a numpy array of booleans.  This
flattens the result in row-major order.  That was the wrong
order for the format of adat['Z'] and adat['UV'] based on the
original matlab data files, around which much of the code had
been built.  As a result, patched issue by transposing all the
arrays so that row-major order was correct.