BUG: Claim iteration number immediately
Previously, an experiment asked for it's iteration number, ran, then
updated the database.  Since multiple concurrent experiments could then
end up with the same iteration, that would be bad.

New behavior is to immediately (though not atomically) insert a trial record
to be updated later with runtime details.  This should be safer behavior
(though not atomic yet).