TST: Add trial2 test
This change adds the test_trials.py test file.

It requires moving some of the parameters of trial2.__main__ to the
global scope, hinting at a more efficient way of creating all tests in
a single file.  In the future, all trials will behave that way

It also manipulates the marks of pytest tests in pyproject.toml and GH
actions

Also: renamed test_integration() test since there's now multiple
integration tests.  May want to remove this in the future