BUG: Specify seed in RigorousParameterSearch2D
Without specifying seed, each n_sim ParameterSearch2D is given the
default seed.  Yes, a default seed was a terrible idea.  But the
paper is almost done, so I'm not going to refactor it out.  If
someone does, it would have to fix the tests as well as the calls
to constructors of ParameterSearch2D.  The latter are hard because
the exp module will detect it as a change in sim parameters,
logging it as incoparable with prior trials.  Which is just a lot
to handle right now.