ENH: Break apart trial and variant in psearch2d CLI
Keeping track of the trial-variant combinations individually is a combinatorial
problem.  So align cli for psearch2d to match that of live_trials, which has
trial and variant passed separately.

Delete all unneccessary trial-variant combo scripts.

This does not move psearch2d to exactly the same syntax as live_trials
(Experiment is part of Variant in psearch2d, but Trial in live_trials)